EnsureTestDevices and EnsureProfiles moved to autoprovision package.
